% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat_tw}
\alias{cat_tw}
\title{Typewriter-style Concatenate and Print}
\usage{
cat_tw(x, sleep = 0.02)
}
\arguments{
\item{x}{A character string of length 1.}

\item{sleep}{Numeric, length 1. System delay in seconds.}
}
\value{
Output to console.
}
\description{
Input text-string and output one character at a time given a delay. Aims to
mimic the Pokemon Blue progressive text reveal, like typewriter output.
}
\examples{
\dontrun{ cat_typewriter("Example text.", 0.2) }
}
