% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safar6-package.R
\docType{package}
\name{safar6-package}
\alias{safar6}
\alias{safar6-package}
\title{safar6: Simulate the Safari Zone from Pokemon Blue with R6}
\description{
An R6-class object to simulate simplified gameplay elements of the
    Safari Zone sub-area from Pokémon Blue (1998) using OOP.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
