% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{SafariZone}
\alias{SafariZone}
\title{R6 Class Representing the Safari Zone from Pokemon Blue (1998)}
\description{
An \code{\link[R6]{R6Class}} object to simulate simplified gameplay elements
of the Safari Zone sub-area from \emph{Pokemon Blue} (1998).
}
\details{
The first generation of \emph{Pokemon} games were developed for the Nintendo
Game Boy by Game Freak (\url{https://www.gamefreak.co.jp/}) and published by
Nintendo (\url{https://www.nintendo.com}). Pokemon as a property is owned by
The Pokemon Company (\url{https://www.pokemon.co.jp/}).
}
\examples{
\dontrun{ x <- safar6::SafariZone$new() }

## ------------------------------------------------
## Method `SafariZone$print`
## ------------------------------------------------

\dontrun{
    x <- safar6::SafariZone  # initialise class
    x$print()  # print the object, see stats
}

## ------------------------------------------------
## Method `SafariZone$pause`
## ------------------------------------------------

\dontrun{
    x <- safar6::SafariZone  # intialise class
    x$pause()  # 'pause' the game, see stats
}

## ------------------------------------------------
## Method `SafariZone$step`
## ------------------------------------------------

\dontrun{
    x <- safar6::SafariZone  # initialise class
    x$step()   # take step, prints steps remaining
}
}
\seealso{
\code{\link{print}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{steps}}{Steps remaining (500 at start).}

\item{\code{balls}}{Safari Balls remaining (30 at start).}

\item{\code{captures}}{Count of wild Pokemon captured (0 at start).}

\item{\code{bills_pc}}{Dataframe of wild Pokemon captured (empty at start).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SafariZone$new()}}
\item \href{#method-print}{\code{SafariZone$print()}}
\item \href{#method-pause}{\code{SafariZone$pause()}}
\item \href{#method-step}{\code{SafariZone$step()}}
\item \href{#method-clone}{\code{SafariZone$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Safari Zone object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SafariZone$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{SafariZone} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Create a new Safari Zone print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SafariZone$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A console message with steps and balls remaining.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
    x <- safar6::SafariZone  # initialise class
    x$print()  # print the object, see stats
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pause"></a>}}
\if{latex}{\out{\hypertarget{method-pause}{}}}
\subsection{Method \code{pause()}}{
Simulate the pause function from the original game.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SafariZone$pause()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A console message with steps and balls remaining.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
    x <- safar6::SafariZone  # intialise class
    x$pause()  # 'pause' the game, see stats
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-step"></a>}}
\if{latex}{\out{\hypertarget{method-step}{}}}
\subsection{Method \code{step()}}{
Take a step in the Safari Zone.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SafariZone$step()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Either nothing, or a wild encounter.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
    x <- safar6::SafariZone  # initialise class
    x$step()   # take step, prints steps remaining
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SafariZone$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
